# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jack Pine Farm Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      description: Returns all non-disabled products (public). Disabled products excluded.
      parameters:
        - in: query
          name: includeDisabled
          schema:
            type: boolean
          description: When true (admin only), include disabled products
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products, admin]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products, admin]
      summary: Update a product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}/notify-me:
    post:
      operationId: subscribeNotifyMe
      tags: [notify-me]
      summary: Subscribe to restock notification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyMeBody"
      responses:
        "200":
          description: Subscribed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/pickup-events:
    get:
      operationId: adminListPickupEvents
      tags: [admin, pickup-events]
      summary: List all pickup events (admin)
      responses:
        "200":
          description: Pickup event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PickupEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreatePickupEvent
      tags: [admin, pickup-events]
      summary: Create a pickup event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePickupEventBody"
      responses:
        "201":
          description: Pickup event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickupEvent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/pickup-events/{id}:
    get:
      operationId: adminGetPickupEvent
      tags: [admin, pickup-events]
      summary: Get pickup event with assigned orders (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pickup event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickupEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdatePickupEvent
      tags: [admin, pickup-events]
      summary: Update a pickup event (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePickupEventBody"
      responses:
        "200":
          description: Pickup event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickupEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/pickup-events/{id}/assign-order:
    post:
      operationId: adminAssignOrderToPickupEvent
      tags: [admin, pickup-events]
      summary: Assign an order to a pickup event (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignOrderBody"
      responses:
        "200":
          description: Order assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/pickup-events/{id}/send-invoices:
    post:
      operationId: adminSendPickupInvoices
      tags: [admin, pickup-events]
      summary: Enter weights and send final balance invoices for a pickup event (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInvoicesBody"
      responses:
        "200":
          description: Invoices sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ProductType:
      type: string
      enum:
        - eggs_chicken
        - eggs_duck
        - meat_chicken
        - meat_turkey
    AvailabilityStatus:
      type: string
      enum:
        - taking_orders
        - preorder
        - sold_out
        - disabled
    PricingType:
      type: string
      enum:
        - unit
        - deposit
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        productType:
          $ref: "#/components/schemas/ProductType"
        pricingType:
          $ref: "#/components/schemas/PricingType"
        priceInCents:
          type: integer
          description: Price in cents. For eggs this is per unit/half-dozen. For meat this is the deposit amount.
        unitLabel:
          type: ["string", "null"]
          description: Display label for the unit (e.g. "dozen", "half-dozen")
        depositDescription:
          type: ["string", "null"]
          description: Description of what the deposit covers (for meat products)
        availability:
          $ref: "#/components/schemas/AvailabilityStatus"
        imageUrl:
          type: ["string", "null"]
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - productType
        - pricingType
        - priceInCents
        - unitLabel
        - depositDescription
        - availability
        - imageUrl
        - displayOrder
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        productType:
          $ref: "#/components/schemas/ProductType"
        pricingType:
          $ref: "#/components/schemas/PricingType"
        priceInCents:
          type: integer
        unitLabel:
          type: ["string", "null"]
        depositDescription:
          type: ["string", "null"]
        availability:
          $ref: "#/components/schemas/AvailabilityStatus"
        imageUrl:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required:
        - name
        - description
        - productType
        - pricingType
        - priceInCents
        - availability
        - displayOrder
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        productType:
          $ref: "#/components/schemas/ProductType"
        pricingType:
          $ref: "#/components/schemas/PricingType"
        priceInCents:
          type: integer
        unitLabel:
          type: ["string", "null"]
        depositDescription:
          type: ["string", "null"]
        availability:
          $ref: "#/components/schemas/AvailabilityStatus"
        imageUrl:
          type: ["string", "null"]
        displayOrder:
          type: integer
    NotifyMeBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
...
    PickupEventStatus:
      type: string
      enum:
        - scheduled
        - completed
        - cancelled
    PickupEvent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scheduledAt:
          type: string
          format: date-time
        locationNotes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/PickupEventStatus"
        assignedOrderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - scheduledAt
        - locationNotes
        - status
        - assignedOrderCount
        - createdAt
        - updatedAt
    CreatePickupEventBody:
      type: object
      properties:
        name:
          type: string
        scheduledAt:
          type: string
          format: date-time
        locationNotes:
          type: ["string", "null"]
      required:
        - name
        - scheduledAt
    UpdatePickupEventBody:
      type: object
      properties:
        name:
          type: string
        scheduledAt:
          type: string
          format: date-time
        locationNotes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/PickupEventStatus"
    AssignOrderBody:
      type: object
      properties:
        orderId:
          type: integer
      required:
        - orderId
    SendInvoicesBody:
      type: object
      properties:
        weights:
          type: array
          items:
            type: object
            properties:
              orderId:
                type: integer
              weightLbs:
                type: number
            required:
              - orderId
              - weightLbs
      required:
        - weights
...