# lib/api-zod/src/generated/types/createPickupEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jack Pine Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePickupEventBody {
  name: string;
  scheduledAt: Date;
  /** @nullable */
  locationNotes?: string | null;
}