# lib/api-zod/src/generated/types/pickupEventStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jack Pine Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export type PickupEventStatus =
  (typeof PickupEventStatus)[keyof typeof PickupEventStatus];

export const PickupEventStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;